+++ workers/tests/processors/test_transcriptome_index.py
@@ -121,7 +121,10 @@ class TXTestCase(TestCase):
         job_context = transcriptome_index._process_gtf(job_context)
 
         transcript_to_gene_ids = pd.read_csv(
-            job_context["genes_to_transcripts_path"], sep="\t", index_col=1, names=["Gene"],
+            job_context["genes_to_transcripts_path"],
+            sep="\t",
+            index_col=1,
+            names=["Gene"],
         )["Gene"]
 
         self.assertEqual(transcript_to_gene_ids.get("ENST00000480647", None), "ENSG00000041988")
