+++ workers/tests/settings.py
@@ -1,0 +1,1 @@
+from data_refinery_workers.settings import *
